genero error al hacer las migraciones

+++ database/migrations/2019_07_22_042139_creacion_tabla_matriculas.php
@@ -18,7 +18,7 @@ class CreacionTablaMatriculas extends Migration
             //Este atributo representa la situación del alumno si es alumno regular(RE), no regular(NR) o egresado(EG).
             $table->string('matriculaSituacion', 2);
             // $table->timestamps();
-            $table->unsignedInteger('estudianteId');
+            $table->unsignedBigInteger('estudianteId');
             $table->foreign('estudianteId')->references('estudianteId')->on('estudiantes');
         });
     }
